Bring Query inspector controls into edit.js
