+++ src/directory/edit.js
@@ -21,6 +21,27 @@ import { useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 
 /**
+ * Inspector Controls appear in the post settings sidebar when a block is being edited.
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/packages/packages-block-editor/#inspectorcontrols
+ */
+import { InspectorControls } from '@wordpress/block-editor';
+
+/**
+ * Core components
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/components/panel/#panelbody
+ */
+import { PanelBody, TextControl } from '@wordpress/components';
+
+/**
+ * Similar string cleanup function to WordPress's `sanitize_title`.
+ * 
+ * @see https://developer.wordpress.org/block-editor/reference-guides/packages/packages-url/#cleanforslug
+ */
+import { cleanForSlug } from '@wordpress/url';
+
+/**
  * Local dependencies
  */
 import SearchForm from "./components/search-form";
@@ -28,6 +49,37 @@ import User from "./components/user";
 import QueryControls from './sidebar/query-controls';
 
 /**
+ * Block controls for the User Query.
+ *
+ * @return {WPElement} Element to render.
+ */
+const QueryInspectorControls = ( { attributes, setAttributes } ) => {
+
+    const { queryId } = attributes;
+
+	return (
+        <InspectorControls>
+            <PanelBody title={__('User Query Settings', 'simple-user-listing')}>
+                <TextControl
+                    label={ __(
+                        'Query ID',
+                        'simple-user-listing'
+                    ) }
+                    help={ __(
+                        'Custom `query_id` for advanced usage.',
+                        'simple-user-listing'
+                    ) }
+                    value={ queryId || '' }
+                    onChange={ ( value ) => {
+                        let cleanValue = cleanForSlug( value );
+                        setAttributes( { queryId: cleanValue } )
+                    } }
+                />
+            </PanelBody>
+        </InspectorControls>
+
+	);
+}
  * The edit function describes the structure of your block in the context of the
  * editor. This represents what the editor will render when the block is used.
  *
@@ -35,7 +87,7 @@ import QueryControls from './sidebar/query-controls';
  *
  * @return {WPElement} Element to render.
  */
-export default function Edit( props ) {
+export default function Edit( { attributes, setAttributes } ) {
     const { users, isLoaded } = useSelect( ( select ) => {
         return {
             users: select( 'core' ).getUsers(),
@@ -45,9 +97,11 @@ export default function Edit( props ) {
 
     return (
         <>
-            <QueryControls {...props} />
+            <QueryInspectorControls
+                attributes={ attributes }
+                setAttributes={ setAttributes }
+            />
 
-	return (
             <div { ...useBlockProps() }>
                 <SearchForm />
                 <div className="user-list-wrap">
